==> notebooks/03_cleaning_categorical_data.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("/content/drive/MyDrive/WMCA/data_cleaning.csv")

# Standardising entries

We will correct recurrent errors in the descriptions and translate/remove the Welsh in the entries.

cat_var = df.select_dtypes(include= ['object','category']).columns.tolist()

# Not these variables
cat_var.remove('address')
cat_var.remove('postcode')
cat_var.remove('msoa_code')
cat_var.remove('lsoa_code')
cat_var.remove('constituency')

# not categorical
cat_var.remove('inspection-date')
cat_var.remove('lodgement-datetime')
cat_var.remove('solar-water-heating-flag')
cat_var.remove('photo-supply-binary')
cat_var.remove('mains-gas-flag')

# remove bilingual 
for c in cat_var:
  df[c] = df[c].str.replace(r"\|(.*)", "")

for c in ['floor-description','walls-description','roof-description']:
    # standardising the unit used
    df[c] = df[c].str.replace(r"W(.*?)K",'W/m²K')
    # removing nonsense 
    df[c] = df[c].str.replace("\*\*\* INVALID INPUT Code \: 57 \*\*\*","")
    # cleaning up 
    df[c] = df[c].str.replace(r'\.$','')
    df[c] = df[c].str.replace(r'\,$','')
    df[c] = df[c].str.replace("  \+"," +")
    df[c] = df[c].str.replace("[ \t]+$","")
    df[c] = df[c].str.replace('  ',' ')
    df[c] = df[c].str.replace('Average thermal transmittance 1 ','Average thermal transmittance 1.00 ')
    df[c] = df[c].str.replace('Average thermal transmittance =','Average thermal transmittance')
    df[c] = df[c].str.rstrip(' ')
    # Lower all case
    df[c] = df[c].str.lower()

# translating welsh sentences
df['walls-description'] = df['walls-description'].str.replace("waliau ceudod","cavity wall")
df['walls-description'] = df['walls-description'].str.replace("dim inswleiddio","no insulation")
df['walls-description'] = df['walls-description'].str.replace("tywodfaen","sandstone")
df['walls-description'] = df['walls-description'].str.replace(r"ceudod wedi(.*?)i lenwi","filled cavity")
df['walls-description'] = df['walls-description'].str.replace(r"wedi(.*?)u hinswleiddio","insulated")
df['walls-description'] = df['walls-description'].str.replace(r"ffr(.*?)m bren","timber frame")
df['walls-description'] = df['walls-description'].str.replace("briciau solet","solid brick")
df['walls-description'] = df['walls-description'].str.replace(r"wedi(.*?)u hadeiladu yn (.*?)l system","system built")
df['walls-description'] = df['walls-description'].str.replace("inswleiddio rhannol","partial insulation")
df['walls-description'] = df['walls-description'].str.replace("gydag inswleiddio allanol","with external insulation")
df['walls-description'] = df['walls-description'].str.replace("gwenithfaen neu risgraig","granite or whinstone")
df['walls-description'] = df['walls-description'].str.replace("gydag inswleiddio mewnol","with internal insulation")
df['walls-description'] = df['walls-description'].str.replace(r"\(rhagdybiaeth\)","(assumed)")
df['walls-description'] = df['walls-description'].str.replace(r"fel y(.*?)u hadeiladwyd, ","")

# removing phrases like as built which aren't adding anything
df['walls-description'] = df['walls-description'].str.replace(r"\(assumed\)","")
df['walls-description'] = df['walls-description'].str.replace("as built, ","")
df['walls-description'] = df['walls-description'].str.replace(", as built","")
df['walls-description'] = df['walls-description'].str.replace(r"cavity\.","cavity wall,")

# standardising punctuation
df['walls-description'] = df['walls-description'].str.replace(r"solid brick\.","solid brick,")

# standardising language
df['walls-description'] = df['walls-description'].replace(
    regex=["granite or whin,", "stone \(granite or whin\)\."],
    value="granite or whinstone,")
df['walls-description'] = df['walls-description'].replace(
    regex=["with external insulation", "with internal insulation", 
     "with additional insulation", "with insulation"],
     value="insulated")

# clean up
df['walls-description'] = df['walls-description'].str.replace(r'\+ chr\(13\) \+','+')
df['walls-description'] = df['walls-description'].str.replace(r"[ \t]+$","")
df['walls-description'] = df['walls-description'].str.replace(r'timber frame\.','timber frame,')
df['walls-description'] = df['walls-description'].str.replace(r'\?','')
df['walls-description'] = df['walls-description'].str.replace(r'system built\.','system built,')

# standardising units
df['roof-description'] = df['roof-description'].str.replace(" mm",'mm')
df['roof-description'] = df['roof-description'].str.replace(">= 300mm",">=300mm")
df['roof-description'] = df['roof-description'].str.replace(">=300mm","300+mm")

# translating welsh sentences
df['roof-description'] = df['roof-description'].str.replace(r"\|\(eiddo arall uwchben\)","")
df['roof-description'] = df['roof-description'].str.replace("ar oleddf","pitched")
df['roof-description'] = df['roof-description'].str.replace("dim inswleiddio","no insulation")
df['roof-description'] = df['roof-description'].str.replace(r"wedi(.*?)i inswleiddio","insulated")
df['roof-description'] = df['roof-description'].str.replace("(rhagdybiaeth)","assumed")
df['roof-description'] = df['roof-description'].str.replace(r"lo inswleiddio yn y llof.*","loft insulation")
df['roof-description'] = df['roof-description'].str.replace(r"o inswleiddio yn y llof.*","loft insulation")
df['roof-description'] = df['roof-description'].str.replace(r"ystafell\(oedd\) to","roof room(s)")
df['roof-description'] = df['roof-description'].str.replace(r"wedi(.*?)i hinswleiddio","insulated")
df['roof-description'] = df['roof-description'].str.replace("nenfwd","ceiling")
df['roof-description'] = df['roof-description'].str.replace(r"wrth y trawstia(.*?)","at rafters")
df['roof-description'] = df['roof-description'].str.replace("inswleiddio cyfyngedig","limited insulation")
df['roof-description'] = df['roof-description'].str.replace("to gwellt, gydag inswleiddio ychwanegol","thatched, with additional insulation")
df['roof-description'] = df['roof-description'].str.replace("yn wastad","always")
df['roof-description'] = df['roof-description'].str.replace("annedd arall uwchben","other premises above")

# standardising descriptions
df['roof-description'] = df['roof-description'].str.replace("roof room,","roof room(s),")
df['roof-description'] = df['roof-description'].str.replace("another dwelling above","other premises above")
df['roof-description'] = df['roof-description'].replace(
    regex=["other premises above","dwelling Above"],
    value="(other premises above)")
df['roof-description'] = df['roof-description'].str.replace("0 w/m²k"," w/m²k")
df['roof-description'] = df['roof-description'].str.replace("  w/m²k"," 0.0 w/m²k")
df['roof-description'] = df['roof-description'].str.replace(r"roof room\(s\), no insulation\(assumed\)","roof room(s), no insulation (assumed)")
df['roof-description'] = df['roof-description'].str.replace(r"\(assumed\)","")
df['roof-description'] = df['roof-description'].str.replace("0mm loft","no")

# cleanup
df['roof-description'] = df['roof-description'].str.replace("mmmm","mm")
df['roof-description'] = df['roof-description'].str.replace(r"\({2,}",r"\(")
df['roof-description'] = df['roof-description'].str.replace(r"\){2,}",r"\)")
df['roof-description'] = df['roof-description'].str.replace(",mm",",")
df['roof-description'] = df['roof-description'].str.replace("thatchedinsulated","thatched, insulated")
df['roof-description'] = df['roof-description'].str.rstrip(" ")

df['roof-description'].value_counts()

# translating welsh sentences
df['floor-description'] = df['floor-description'].str.replace(r"anheddiad arall islaw",'another dwelling below') 
df['floor-description'] = df['floor-description'].str.replace(r"\(eiddo arall islaw\)",'other premises below') 
df['floor-description'] = df['floor-description'].str.replace(r"wedi(.*?)i inswleiddio","insulated")
df['floor-description'] = df['floor-description'].str.replace("dim inswleiddio","no insulation")
df['floor-description'] = df['floor-description'].str.replace("(rhagdybiaeth)","assumed")
df['floor-description'] = df['floor-description'].str.replace("crog","suspended")
df['floor-description'] = df['floor-description'].str.replace("heb ei inswleiddio","no insulation")
df['floor-description'] = df['floor-description'].str.replace("i ofod heb ei wresogi","to unheated space")
df['floor-description'] = df['floor-description'].str.replace("solet","solid")
df['floor-description'] = df['floor-description'].str.replace("inswleiddio cyfyngedig","limited insulation")

# removing assumed
df['floor-description'] = df['floor-description'].str.replace(r"\(assumed\)","")

# standardising language
df['floor-description'] = df['floor-description'].replace(
    regex=["\(another dwelling below\)", "\(other premises below\)"],
    value='other premises below') 
df['floor-description'] = df['floor-description'].str.replace(r"solid\.",'solid,') 
df['floor-description'] = df['floor-description'].str.replace("uninsulated",'no insulation,') 
df['floor-description'] = df['floor-description'].str.replace("insulation=100mm",'100 mm insulation') 
df['floor-description'] = df['floor-description'].str.replace(r"\, \(assumed\)",' (assumed)')
df['floor-description'] = df['floor-description'].str.replace("insulation=25mm",'25 mm insulation')
df['floor-description'] = df['floor-description'].str.replace("insulation=75mm",'75 mm insulation')
df['floor-description'] = df['floor-description'].str.replace("limited insulated",'limited insulation')
df['floor-description'] = df['floor-description'].str.replace("(same dwelling below) ", "")
# cleanup
df['floor-description'] = df['floor-description'].str.replace(r'^, ','')
df['floor-description'] = df['floor-description'].str.replace(r'\?','')
df['floor-description'] = df['floor-description'].str.rstrip(' ,.')

# Extract thermal transmittance feature
##improve null##
for c in ['floor-description','walls-description','roof-description']:
  col_name = c.split('-')[0] + "-thermal-transmittance"
  transmittance_col = df[c].str.findall(r'(\d.\d*) W/m²K')
  df[col_name] = round(transmittance_col.str[0].astype(float),2)

construction_year = df['construction-age-band'].str.findall(r"(\d{4})")
df['construction-age-band'] = construction_year.str[-1].astype('datetime64')

cat_var.remove('construction-age-band')

df['transaction-type'] = df['transaction-type'].str.replace(" - this is for backwards compatibility only and should not be used","")
df['transaction-type'] = df['transaction-type'].str.replace("Stock Condition Survey", "Stock condition survey")
df['transaction-type'] = df['transaction-type'].replace("not recorded", np.nan)

# standardising language
df['main-fuel'] = df['main-fuel'].str.replace('Gas: ', '')
df['main-fuel'] = df['main-fuel'].str.replace(r'Electricity\:.*', 'electricity')

# cleanup
df['main-fuel'] = df['main-fuel'].str.replace(" - this is for backwards compatibility only and should not be used","")
df['main-fuel'] = df['main-fuel'].replace('To be used only when there is no heating/hot-water system', np.nan)
df['main-fuel'] = df['main-fuel'].replace('To be used only when there is no heating/hot-water system or data is from a community network', np.nan)

df['tenure'] = df['tenure'].str.replace('Rented', 'rental')
df['tenure'] = df['tenure'].replace('Not defined - use in the case of a new dwelling for which the intended tenure in not known. It is no', np.nan)
df['tenure'] = df['tenure'].str.lower()

df['floor-level'] = df['floor-level'].replace(
    regex=[r'(?<=\d)th', r'(?<=\d)nd', r'(?<=\d)rd', r'(?<=\d)st'], 
    value='')
df['floor-level'] = df['floor-level'].str.lstrip('0')
df['floor-level'] = df['floor-level'].replace(
    regex=['ground floor', r'^\s*$'], 
    value='Ground')
df['floor-level'] = df['floor-level'].replace('Basement', '-1')

df['glazed-area'] = df['glazed-area'].replace('Not Defined', np.nan)

df['mainheatcont-description'] = df['mainheatcont-description'].str.lower()

# translating welsh sentences
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace(r"rheoli.r t.l . llaw","manual charge control")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("rhaglennydd, dim thermostat ystafell","programmer, no room thermostat")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("rheolaeth amser a rheolaeth parthau tymheredd","time and temperature zone control")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("rhaglennydd a thermostat ystafell","programmer and room thermostat")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("rhaglennydd a thermostatau ar y cyfarpar","programmer and appliance thermostats")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("rhaglennydd ac o leiaf ddau thermostat ystafell","programmer and at least two room thermostats")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("thermostat ystafell yn unig","room thermostat only")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("dim rheolaeth thermostatig ar dymheredd yr ystafell","no thermostatic control of room temperature")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("rheoli gwefr drydanol yn awtomatig","automatic charge control")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("dim rheolaeth amser na rheolaeth thermostatig ar dymheredd yr ystafell","no time or thermostatic control of room temperature")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("trvs a falf osgoi","trvs and bypass")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("rhaglennydd","programmer")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("tal un gyfradd, thermostat ystafell yn unig","flat rate charging, room thermostat only")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("thermostat ystafell a trvs","room thermostat and trvs")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace(r"t(.*?)l un gyfradd","flat rate charging")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("thermostatau ar y cyfarpar","appliance thermostat")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("dim","none")

# standardising language
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("programmer, no thermostat","programmer, no room thermostat")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("flat rate charging\*","flat rate charging")
df['mainheatcont-description'] = df['mainheatcont-description'].replace(
    regex=[r"\+",r"\&"] ,
    value="and")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("trv.s","trvs")
df['mainheatcont-description'] = df['mainheatcont-description'].replace(
    regex=["thermostats", " stat", "thermostatic"],
    value="thermostat")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("communit ","community ")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("to the use of community heating","to use of community heating")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("controls","control")
df['mainheatcont-description'] = df['mainheatcont-description'].replace(
    regex=["prog ","program "],
    value="programmer ")
df['mainheatcont-description'] = df['mainheatcont-description'].replace(
    regex=["programmerand",r"programmer\?and"],
    value="programmer and")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("roomthermostat","room thermostat")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("delayed start thermostat and program and trvs","delayed start thermostat, program and trvs")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("flat rate charging, programmer no room thermostat","flat rate charging, programmer, no room thermostat")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace(" 2 "," two ")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("roomstat","room thermostat")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("programmer and room thermostat and trvs","programmer, room thermostat and trvs")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("programmer and trvs and boiler energy manager","programmer, trvs and boiler energy manager")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("programmer and trvs and bypass","programmer, trvs and bypass")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("programmer and trvs and flow switch","programmer, trvs and flow switch")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace(r"temp+$","temperature")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("no thermostat control of room temperature","no thermostat control")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("appliance thermostat and programmer","programmer and appliance thermostat")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("delayed start thermostat and programmer and trvs","delayed start thermostat, programmer and trvs")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("no time or thermostat control of temperature","no time or thermostat control of room temperature")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("nothermostat","no thermostat")
df['mainheatcont-description'] = df['mainheatcont-description'].str.replace("at least two room thermostat","at least two room thermostats")

# translating welsh sentences
df['hotwater-description'] = df['hotwater-description'].str.replace(r"O(.*?)r brif system","From main system")
df['hotwater-description'] = df['hotwater-description'].str.replace(r"Trochi trydan","Electric immersion")
df['hotwater-description'] = df['hotwater-description'].str.replace(r"an-frig","off-peak")
df['hotwater-description'] = df['hotwater-description'].str.replace(r"O system eilaidd","From secondary system")
df['hotwater-description'] = df['hotwater-description'].str.replace(r"Nwy wrth fwy nag un pwynt","Gas multipoint")
df['hotwater-description'] = df['hotwater-description'].str.replace(r"Popty estynedig olew","Oil range cooker")
df['hotwater-description'] = df['hotwater-description'].str.replace(r"tarriff safonol","standard tariff")
df['hotwater-description'] = df['hotwater-description'].str.replace(r"Dim system ar gael","No system present")
df['hotwater-description'] = df['hotwater-description'].str.replace(r"adfer gwres nwyon ffliw","flue gas heat recovery")
df['hotwater-description'] = df['hotwater-description'].str.replace(r"gydag ynni(.*?)r haul","plus solar")
df['hotwater-description'] = df['hotwater-description'].str.replace(r"dim thermostat ar y silindr","no cylinderstat")
df['hotwater-description'] = df['hotwater-description'].str.replace(r"rhagdybir bod twymwr tanddwr trydan","electric immersion assumed")
df['hotwater-description'] = df['hotwater-description'].str.replace(r"an-frig","off peak")
df['hotwater-description'] = df['hotwater-description'].str.replace(r"Twymwr tanddwr","underfloor heating")

# Standardising language
df['hotwater-description'] = df['hotwater-description'].str.replace("cylinder thermostat","cylinderstat")
df['hotwater-description'] = df['hotwater-description'].replace(
    regex=["No system present :", "No hot water system present -"],
    value="No system present:")
df['hotwater-description'] = df['hotwater-description'].replace(
    regex=["From community scheme", "community scheme"],
    value="Community scheme")
df['hotwater-description'] = df['hotwater-description'].str.replace("Community heat pump","Community scheme with CHP")
df['hotwater-description'] = df['hotwater-description'].replace(
    regex=["From secondary heater","From second main heating system"],
    value="From secondary system")
df['hotwater-description'] = df['hotwater-description'].str.replace("SAP05:Hot-Water","SAP:Hot-Water")
df['hotwater-description'] = df['hotwater-description'].str.replace("plus solar, no cylinderstat","no cylinderstat, plus solar")
df['hotwater-description'] = df['hotwater-description'].str.replace("none","No system present: electric immersion assumed")
df['hotwater-description'] = df['hotwater-description'].str.replace("plus solar, flue gas heat recovery","flue gas heat recovery, plus solar")
df['hotwater-description'] = df['hotwater-description'].str.replace("no cylinderstat, no cylinderstat","no cylinderstat")

# cleanup
df['hotwater-description'] = df['hotwater-description'].replace("***SAMPLE***",np.nan)
df['hotwater-description'] = df['hotwater-description'].str.strip(" .,")

df['windows-description'] = df['windows-description'].str.lower()

# translating welsh sentences
df['windows-description'] = df['windows-description'].str.replace("ffenestri perfformiad uchel","high performance glazing")
df['windows-description'] = df['windows-description'].str.replace("gwydrau dwbl gan mwyaf","mostly double glazing")
df['windows-description'] = df['windows-description'].str.replace("rhai gwydrau dwbl","partial double glazing")
df['windows-description'] = df['windows-description'].str.replace("gwydrau sengl","single glazing")
df['windows-description'] = df['windows-description'].str.replace("gwydrau dwbl rhannol","partial double glazing")
df['windows-description'] = df['windows-description'].str.replace("gwydrau dwbl llawn","fully double glazing")
df['windows-description'] = df['windows-description'].str.replace("gwydrau lluosog ym mhobman","multiple glazing throughout")
df['windows-description'] = df['windows-description'].str.replace("gwydrau eilaidd llawn","full secondary glazing")

# Standardising language
df['windows-description'] = df['windows-description'].str.replace("glazed","glazing")
df['windows-description'] = df['windows-description'].str.replace("fully","full")

# clean-up
df['windows-description'] = df['windows-description'].str.replace("single glazingsingle glazing","single glazing")
df['windows-description'] = df['windows-description'].str.replace("single glazingdouble glazing","single glazing and double glazing")
df['windows-description'] = df['windows-description'].str.replace("single glazingsecondary glazing","single glazing and secondary glazing")
df['windows-description'] = df['windows-description'].str.replace("  "," ")

df['secondheat-description1'] = df.apply(lambda row: row['lighting-description'] if 'lighting' in str(row['secondheat-description']) else row['secondheat-description'],axis = 1)
df['lighting-description1'] = df.apply(lambda row: row['secondheat-description'] if 'lighting' in str(row['secondheat-description']) else row['lighting-description'],axis = 1)
df.drop(columns= ['secondheat-description','lighting-description'],axis=1,inplace=True)
df.rename(columns={'secondheat-description1':'secondheat-description','lighting-description1':'lighting-description'},inplace=True)

## CONVERT TO NUMERICAL ##
df['lighting-description'] = df['lighting-description'].str.lower()

# Converting lighting descriptions in % of low energy lighting
df['lighting-description'] = df['lighting-description'].str.replace(r'goleuadau ynni-isel ym mhob un o.r mannau gosod','low energy lighting in 100% fixed outlets')
df['lighting-description'] = df['lighting-description'].str.replace(r'low energy lighting in 120% of fixed outlets','low energy lighting in 100% fixed outlets')
df['lighting-description'] = df['lighting-description'].str.replace(r'low energy lighting 100% of fixed outlets','low energy lighting in 100% fixed outlets')
df['lighting-description'] = df['lighting-description'].str.replace(r'no low energy lighting','low energy lighting in 0% fixed outlets')
df['lighting-description'] = df['lighting-description'].str.replace('dim goleuadau ynni-isel', 'low energy lighting in 0% fixed outlets')

# Extract % low energy lighting
df['percentage-low-energy-lighting'] = df['lighting-description'].str.extract('([-+]?(?:\d*\.\d+|\d+))')
df['percentage-low-energy-lighting'] = df['percentage-low-energy-lighting'].astype(float)

# standardise language
df['secondheat-description'] = df['secondheat-description'].str.rstrip(', ')
df['secondheat-description'] = df['secondheat-description'].str.replace('LNG', 'LPG')
df['secondheat-description'].replace(regex=[r'\,\s\(null\)', r'\s*\(assumed\)'], value='', inplace=True)

# translate welsh
df['secondheat-description'] = df['secondheat-description'].str.replace('Gwresogyddion ystafell, nwy prif gyflenwad', 'Room heaters, mains gas')
df['secondheat-description'] = df['secondheat-description'].str.replace('Gwresogyddion ystafell, trydan', 'Room heaters, electric')
df['secondheat-description'] = df['secondheat-description'].str.replace('Gwresogyddion trydan cludadwy', 'Portable electric heaters')

df['secondheat-description'] = df['secondheat-description'].str.lower()

df['main-heating-controls'] = df['main-heating-controls'].replace('%%MAINHEATCONTROL%%', np.nan)
df['main-heating-controls'] = df['main-heating-controls'].astype(float)

cat_var.remove('main-heating-controls')

# check that cleaning is done correctly
for c in cat_var:
  print(f'\nVariable: {c}')
  print(df[c].unique())

df.to_csv("/content/drive/MyDrive/WMCA/cleaning_categorical_data.csv", index=False)
