==> notebooks/05_encoding_categorical.ipynb <==
import pandas as pd
import numpy as np
import datetime as dt

from sklearn.preprocessing import LabelEncoder
from sklearn_pandas import CategoricalImputer
from sklearn.preprocessing import OneHotEncoder

df = pd.read_csv("/content/drive/MyDrive/WMCA/chaid_data.csv")
df.drop(columns=["current-energy-rating", "current-energy-efficiency",
                'address'], inplace=True)

# Convert to ordinal for regression
df['inspection-date'] = df['inspection-date'].astype('datetime64').map(dt.datetime.toordinal)
df['lodgement-datetime'] = df['lodgement-datetime'].astype('datetime64').map(dt.datetime.toordinal)

# Keep missing values as seperate category
bool_var = ['solar-water-heating-flag', 'photo-supply-binary', 'mains-gas-flag']
df[bool_var] = df[bool_var].replace(np.nan, 'MISSING')
df[bool_var] = df[bool_var].astype(str)

# One hot encode non-ordinal variables
for var in bool_var:
  enc = OneHotEncoder(handle_unknown='ignore')
  df[var] = enc.fit_transform(np.array(df[var]).reshape(-1,1)).toarray()

Impute with iterator (merge with other)

cat_var = df.select_dtypes(include= ['object']).columns.tolist()

percent_missing = df[cat_var].isnull().sum() * 100 / len(df)
percent_missing.sort_values(ascending=False, inplace=True)
print(percent_missing)

# High missing values 
df.drop(columns=['floor-level'], inplace=True)
cat_var.remove('floor-level')

# Impute small amount of missing
for var in percent_missing.index[percent_missing <1]:
  imputer = CategoricalImputer()
  df[var] = imputer.fit_transform(df[var])

# Fill remaining missing values
df[cat_var] = df[cat_var].replace(np.nan, 'MISSING')

ordinal_var = [col for col in df.columns if '-eff' in col] + ['construction-age-band', 'glazed-area']

# Change categorical variable to numbers
for var in ordinal_var:
  label_encoder = LabelEncoder()
  df[var] = label_encoder.fit_transform(df[var]) 

non_ordinal_var = [col for col in df.columns if col not in ordinal_var]
# One hot encode non-ordinal variables
for var in non_ordinal_var:
  # Crashes :(
  # enc = OneHotEncoder(handle_unknown='ignore')
  # df[var] = enc.fit_transform(np.array(df[var]).reshape(-1,1)).toarray()

  label_encoder = LabelEncoder()
  df[var] = label_encoder.fit_transform(df[var]) 

df.to_csv("/content/drive/MyDrive/WMCA/encoding_categorical.csv", index=False)
